+++ packages/icon/src/lib/Icons/Actions.tsx
@@ -21,6 +21,7 @@ import { faPencil } from '@fortawesome/free-solid-svg-icons/faPencil';
 import { faPlay } from '@fortawesome/free-solid-svg-icons/faPlay';
 import { faPlus } from '@fortawesome/free-solid-svg-icons/faPlus';
 import { faCloudArrowDown } from '@fortawesome/free-solid-svg-icons/faCloudArrowDown';
+import { faCloudArrowUp } from '@fortawesome/free-solid-svg-icons/faCloudArrowUp';
 import { faPlusCircle } from '@fortawesome/free-solid-svg-icons/faPlusCircle';
 import { faTag } from '@fortawesome/free-solid-svg-icons/faTag';
 import { faTags } from '@fortawesome/free-solid-svg-icons/faTags';
@@ -95,3 +96,5 @@ export const PauseIcon = (props: SvgIconProps): JSX.Element => <FaSvgIcon icon={
 export const GearsIcon = (props: SvgIconProps): JSX.Element => <FaSvgIcon icon={faGears} {...props} />;
 
 export const CloudDownloadIcon = (props: SvgIconProps): JSX.Element => <FaSvgIcon icon={faCloudArrowDown} {...props} />;
+
+export const CloudUploadIcon = (props: SvgIconProps): JSX.Element => <FaSvgIcon icon={faCloudArrowUp} {...props} />;