+++ packages/theme/src/lib/light-theme.ts
@@ -74,8 +74,11 @@ export const lightTheme = {
       inputFocus: tokens.borderInputFocus,
       inverse: tokens.borderInverse,
       inverseSecondary: tokens.borderInverseSecondary,
+      hover: tokens.colorInfoMain20,
     },
     background: {
+      primary: tokens.colorPrimaryMain20,
+      secondary: '#393B4621',
       canvas: tokens.colorBackgroundCanvas,
       paper: tokens.colorBackgroundPaper,
       header: tokens.colorBackgroundHeader,
@@ -434,7 +437,7 @@ export const lightTheme = {
           '.MuiListItemButton-root': {
             borderRadius: '0 4px 4px 0',
             '&:active': {
-              backgroundColor: tokens.colorActionFocus
+              backgroundColor: tokens.colorActionFocus,
             },
             '&.Mui-focusVisible': {
               border: `2px solid ${tokens.colorPrimaryMain}`,
@@ -540,11 +543,12 @@ export const lightTheme = {
               '.MuiAccordionSummary-root:not(.Mui-focusVisible, :hover, :active)': {
                 backgroundColor: 'inherit',
               },
-              '> .MuiCollapse-root > .MuiCollapse-wrapper > .MuiCollapse-wrapperInner > .MuiAccordion-region > .MuiAccordionDetails-root:first-of-type': {
-                borderTop: 'none'
-              }
+              '> .MuiCollapse-root > .MuiCollapse-wrapper > .MuiCollapse-wrapperInner > .MuiAccordion-region > .MuiAccordionDetails-root:first-of-type':
+                {
+                  borderTop: 'none',
+                },
             },
-          }
+          },
         },
       },
     },
@@ -552,18 +556,18 @@ export const lightTheme = {
       styleOverrides: {
         root: {
           padding: '1.25rem',
-          flexWrap: 'wrap'
+          flexWrap: 'wrap',
         },
         spacing: {
-          rowGap: tokens.spacingBase
-        }
+          rowGap: tokens.spacingBase,
+        },
       },
     },
     MuiAccordionDetails: {
       styleOverrides: {
         root: {
           minHeight: '1px',
-          padding: '1.25rem'
+          padding: '1.25rem',
         },
       },
     },
@@ -595,7 +599,7 @@ export const lightTheme = {
           order: 2,
           fontWeight: tokens.fontWeightsBold,
           margin: '12px 0px',
-          gap: 1
+          gap: 1,
         },
         expandIconWrapper: {
           order: 1,
@@ -624,11 +628,11 @@ export const lightTheme = {
             margin: '0px -4px',
           },
           '.MuiIconButton-root': {
-            marginRight: '-8px'
+            marginRight: '-8px',
           },
           margin: '-2px 0px -2px auto',
           padding: '0 0 0 16px',
-          gap: '16px'
+          gap: '16px',
         },
         icon: ({ ownerState }: { ownerState: any }) => ({
           ...(ownerState.severity === 'success' && {
@@ -713,14 +717,14 @@ export const lightTheme = {
             height: '2.25rem',
             '.MuiSelect-avExpandIcon': {
               transform: 'unset',
-              top: 'auto'
+              top: 'auto',
             },
           },
         },
         popupIndicator: {
           color: tokens.borderInput,
 
-          padding: '.75rem'
+          padding: '.75rem',
         },
         popupIndicatorOpen: {
           transform: 'scaleY(-1)',
@@ -977,7 +981,7 @@ export const lightTheme = {
         sizeSmall: {
           verticalAlign: 'text-bottom',
           borderRadius: '50px',
-          height: 'min-content'
+          height: 'min-content',
         },
         sizeMedium: {
           fontWeight: tokens.fontWeightsRegular,
@@ -995,9 +999,9 @@ export const lightTheme = {
     MuiDateCalendar: {
       styleOverrides: {
         root: {
-          width: '310px'
-        }
-      }
+          width: '310px',
+        },
+      },
     },
     MuiDatePicker: {
       styleOverrides: {
@@ -1017,42 +1021,55 @@ export const lightTheme = {
       },
     },
     MuiDialog: {
+      defaultProps: {
+        PaperProps: {
+          variant: 'elevation',
+          elevation: 4,
+        },
+      },
       styleOverrides: {
         maxWidthSm: {
-          maxWidth: '25rem'
+          maxWidth: '25rem',
         },
         maxWidthMd: {
-          maxWidth: '39.75rem'
+          maxWidth: '39.75rem',
         },
         maxWidthLg: {
-          maxWidth: '58.75rem'
+          maxWidth: '58.75rem',
         },
         maxWidthXl: {
-          maxWidth: '71.25rem'
+          maxWidth: '71.25rem',
         },
         avCloseButton: {
-          "~ .MuiDialogTitle-root": {
-            marginRight: '2.5rem'
-          }
-        }
-      }
+          '~ .MuiDialogTitle-root': {
+            marginRight: '2.5rem',
+          },
+        },
+      },
     },
     MuiDialogActions: {
       styleOverrides: {
         root: {
-          padding: '1rem'
-        }
-      }
+          padding: '1rem',
+          flexWrap: 'wrap',
+        },
+        spacing: {
+          marginBottom: '-8px',
+          '.MuiButton-root': {
+            marginBottom: '8px',
+          },
+        },
+      },
     },
     MuiDialogContent: {
       defaultProps: {
-        dividers: true
+        dividers: true,
       },
       styleOverrides: {
         root: {
-          padding: '1rem'
-        }
-      }
+          padding: '1rem',
+        },
+      },
     },
     MuiDialogTitle: {
       styleOverrides: {
@@ -1061,9 +1078,9 @@ export const lightTheme = {
           paddingLeft: '1rem',
           minHeight: '3.5rem',
           display: 'flex',
-          alignItems: 'center'
-        }
-      }
+          alignItems: 'center',
+        },
+      },
     },
     MuiFormControl: {
       defaultProps: {
@@ -1709,12 +1726,14 @@ export const lightTheme = {
             // full width
             left: '24px',
             right: '24px',
-            '&.MuiSnackbar-anchorOriginBottomCenter, &.MuiSnackbar-anchorOriginBottomLeft, &.MuiSnackbar-anchorOriginBottomRight': {
-              bottom: '24px',
-            },
-            '&.MuiSnackbar-anchorOriginTopCenter, &.MuiSnackbar-anchorOriginTopLeft, &.MuiSnackbar-anchorOriginTopRight': {
-              top: '24px',
-            },
+            '&.MuiSnackbar-anchorOriginBottomCenter, &.MuiSnackbar-anchorOriginBottomLeft, &.MuiSnackbar-anchorOriginBottomRight':
+              {
+                bottom: '24px',
+              },
+            '&.MuiSnackbar-anchorOriginTopCenter, &.MuiSnackbar-anchorOriginTopLeft, &.MuiSnackbar-anchorOriginTopRight':
+              {
+                top: '24px',
+              },
           },
           [theme.breakpoints.up('sm')]: {
             maxWidth: '528px',
@@ -1725,7 +1744,7 @@ export const lightTheme = {
           },
           '.MuiAlert-root, .MuiSnackbarContent-root': {
             boxShadow: tokens.shadows24,
-          }
+          },
         }),
         anchorOriginBottomCenter: {
           bottom: '40px',
@@ -1756,9 +1775,9 @@ export const lightTheme = {
         root: {
           // info alert styling
           backgroundColor: tokens.colorBlue50,
-          color: tokens.colorTextPrimary
-        }
-      }
+          color: tokens.colorTextPrimary,
+        },
+      },
     },
     MuiStep: {
       styleOverrides: {
@@ -1941,6 +1960,95 @@ export const lightTheme = {
         },
         fontSizeXlarge: {
           fontSize: '2.25rem',
+        },
+      },
+    },
+    MuiSwitch: {
+      defaultProps: {
+        disableRipple: true,
+        disableFocusRipple: true,
+        disableTouchRipple: true,
+      },
+      styleOverrides: {
+        root: {
+          padding: 'unset',
+          height: '20px',
+          width: '40px',
+          overflow: 'initial',
+          '.MuiSwitch-switchBase': {
+            padding: '2px',
+            '&.Mui-focusVisible': {
+              '+ .MuiSwitch-track': {
+                outlineWidth: tokens.borderWidthXs,
+                outlineColor: tokens.colorTextInverse,
+                outlineStyle: 'solid',
+                boxShadow: `0 0 0px 4px ${tokens.borderPrimary}`,
+              },
+            },
+            '&:hover': {
+              '+ .MuiSwitch-track': {
+                backgroundColor: tokens.colorSecondaryMain,
+              },
+            },
+            '&.Mui-checked': {
+              ' + .MuiSwitch-track': {
+                backgroundColor: tokens.colorPrimaryLight,
+                opacity: 1,
+              },
+              '&::hover': {
+                '+ .MuiSwitch-track': {
+                  backgroundColor: tokens.colorPrimaryDark,
+                },
+              },
+            },
+            '&.Mui-disabled, &.Mui-disabled.Mui-checked': {
+              color: tokens.colorCommonWhite,
+              '+ .MuiSwitch-track': {
+                backgroundColor: tokens.colorTextDisabled,
+                opacity: 1,
+              },
+            },
+          },
+        },
+        thumb: {
+          height: '16px',
+          width: '16px',
+          boxShadow: 'unset',
+          backgroundColor: tokens.colorCommonWhite,
+        },
+        track: {
+          borderRadius: '10px',
+          background: tokens.colorSecondaryLight,
+          opacity: 'unset',
+        },
+        sizeSmall: {
+          height: '16px',
+          width: '32px',
+          '.MuiSwitch-switchBase': {
+            '&.Mui-checked': {
+              transform: 'translateX(16px)',
+            },
+          },
+          '.MuiSwitch-thumb': {
+            height: '12px',
+            width: '12px',
+          },
+          '.MuiSwitch-track': {
+            borderRadius: '8px',
+          },
+        },
+      },
+    },
+    AvSwitchCheckedIcon: {
+      styleOverrides: {
+        root: {
+          '& .Mui-checked': {
+            '+ .MuiSwitch-track': {
+              '&::before': {
+                left: 4,
+              },
+            },
+          },
         },
       },
     },
@@ -2167,18 +2275,18 @@ export const lightTheme = {
       styleOverrides: {
         checkbox: {
           '&.Mui-checked': {
-            color: tokens.colorTextInverse
-          }
+            color: tokens.colorTextInverse,
+          },
         },
         content: {
           borderRadius: 0,
           padding: '8px 16px',
           alignItems: 'start',
           '&:hover': {
-            backgroundColor: tokens.colorActionHover
+            backgroundColor: tokens.colorActionHover,
           },
           '&:active': {
-            backgroundColor: `${tokens.colorSecondaryDark}26`
+            backgroundColor: `${tokens.colorSecondaryDark}26`,
           },
           '&.Mui-focused': {
             outlineWidth: tokens.borderWidthXs,
@@ -2199,18 +2307,18 @@ export const lightTheme = {
             },
             '&:active': {
               backgroundColor: tokens.colorPrimaryDarker,
-            }
+            },
           },
         },
         iconContainer: {
           color: tokens.colorSecondaryMain,
           paddingTop: '4px',
-          'svg': {
+          svg: {
             fontSize: 'inherit',
-            marginRight: '4px'
-          }
+            marginRight: '4px',
+          },
         },
-      }
+      },
     },
     MuiTypography: {
       defaultProps: {