+++ packages/favorites/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.3](https://github.com/Availity/element/compare/@availity/mui-favorites@0.2.2...@availity/mui-favorites@0.2.3) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.2.2`
+* `mui-progress` updated to version `0.2.2`
+## [0.2.2](https://github.com/Availity/element/compare/@availity/mui-favorites@0.2.1...@availity/mui-favorites@0.2.2) (2025-01-03)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.2.1`
+* `mui-progress` updated to version `0.2.1`
 ## [0.2.1](https://github.com/Availity/element/compare/@availity/mui-favorites@0.2.0...@availity/mui-favorites@0.2.1) (2024-12-09)
 
 ### Dependency Updates