+++ packages/feedback/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.5](https://github.com/Availity/element/compare/@availity/mui-feedback@0.3.4...@availity/mui-feedback@0.3.5) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-button` updated to version `0.3.4`
+* `mui-form-utils` updated to version `0.3.4`
+* `mui-icon` updated to version `0.3.4`
+* `mui-textfield` updated to version `0.3.4`
+* `mui-toggle-button` updated to version `0.3.4`
+## [0.3.4](https://github.com/Availity/element/compare/@availity/mui-feedback@0.3.3...@availity/mui-feedback@0.3.4) (2025-01-03)
+
+### Dependency Updates
+
+* `mui-button` updated to version `0.3.3`
+* `mui-form-utils` updated to version `0.3.3`
+* `mui-icon` updated to version `0.3.3`
+* `mui-textfield` updated to version `0.3.3`
+* `mui-toggle-button` updated to version `0.3.3`
+## [0.3.3](https://github.com/Availity/element/compare/@availity/mui-feedback@0.3.2...@availity/mui-feedback@0.3.3) (2024-12-23)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `0.3.2`
+* `mui-textfield` updated to version `0.3.2`
 ## [0.3.2](https://github.com/Availity/element/compare/@availity/mui-feedback@0.3.1...@availity/mui-feedback@0.3.2) (2024-12-09)
 
 ### Dependency Updates