+++ packages/progress/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.4.3](https://github.com/Availity/element/compare/@availity/mui-progress@0.4.2...@availity/mui-progress@0.4.3) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.4.2`
+## [0.4.2](https://github.com/Availity/element/compare/@availity/mui-progress@0.4.1...@availity/mui-progress@0.4.2) (2025-01-03)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.4.1`
 ## [0.4.1](https://github.com/Availity/element/compare/@availity/mui-progress@0.4.0...@availity/mui-progress@0.4.1) (2024-12-09)
 
 ### Dependency Updates