+++ packages/stepper/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.4](https://github.com/Availity/element/compare/@availity/mui-stepper@0.2.3...@availity/mui-stepper@0.2.4) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-button` updated to version `0.2.3`
+* `mui-icon` updated to version `0.2.3`
+## [0.2.3](https://github.com/Availity/element/compare/@availity/mui-stepper@0.2.2...@availity/mui-stepper@0.2.3) (2025-01-03)
+
+### Dependency Updates
+
+* `mui-button` updated to version `0.2.2`
+* `mui-icon` updated to version `0.2.2`
 ## [0.2.2](https://github.com/Availity/element/compare/@availity/mui-stepper@0.2.1...@availity/mui-stepper@0.2.2) (2024-12-09)
 
 ### Dependency Updates