+++ packages/textfield/CHANGELOG.md
@@ -2,6 +2,29 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.6.16](https://github.com/Availity/element/compare/@availity/mui-textfield@0.6.15...@availity/mui-textfield@0.6.16) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `0.6.15`
+* `mui-icon` updated to version `0.6.15`
+* `mui-button` updated to version `0.6.15`
+* `mui-chip` updated to version `0.6.15`
+* `mui-menu` updated to version `0.6.15`
+## [0.6.15](https://github.com/Availity/element/compare/@availity/mui-textfield@0.6.14...@availity/mui-textfield@0.6.15) (2025-01-03)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `0.6.14`
+* `mui-icon` updated to version `0.6.14`
+* `mui-button` updated to version `0.6.14`
+* `mui-chip` updated to version `0.6.14`
+* `mui-menu` updated to version `0.6.14`
+## [0.6.14](https://github.com/Availity/element/compare/@availity/mui-textfield@0.6.13...@availity/mui-textfield@0.6.14) (2024-12-23)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `0.6.13`
 ## [0.6.13](https://github.com/Availity/element/compare/@availity/mui-textfield@0.6.12...@availity/mui-textfield@0.6.13) (2024-12-09)
 
 ### Dependency Updates