+++ packages/tree/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.6](https://github.com/Availity/element/compare/@availity/mui-tree@0.2.5...@availity/mui-tree@0.2.6) (2025-01-06)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.2.5`
+## [0.2.5](https://github.com/Availity/element/compare/@availity/mui-tree@0.2.4...@availity/mui-tree@0.2.5) (2025-01-03)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.2.4`
 ## [0.2.4](https://github.com/Availity/element/compare/@availity/mui-tree@0.2.3...@availity/mui-tree@0.2.4) (2024-12-09)
 
 ### Dependency Updates